Pool puppeteer jobs (#11)
Scrape jobs are now in a pool. While scraping the main/search pages a job queue(job pool) is build. Once a resource in the pool is free a new job is started. This reducer the time to scrape multible pages about the same amount as the max queue defined. Currently the queue is maxed at 5 active jobs. This seams to be a god value as GitHub actions VM's has about 2GHz of CPU to run this task; Any more concurrent active jobs slows the overall performance.

Minor:
* Action `.github/workflows/scrape-all-to-pr.js.yml` updated to scrape each discount service as a job matrix

* Moved utility functions to its own folder

* Extended som of the utility functions

* Start up questions when running index.js now generates more fleksible

* Introduced a watchdog time to kill the process if it runs too long